break shells up into separate directory
